Create molecular_orbittals
This file first relax the structure and then find out the number of HOMO and LUMO orbitals. Create a new directory MO in current directory and then submit the jobs to create PARCHRG file for HOMO and LUMO. The corresponding PARCHARG files will be available inside the MO/HOMO and MO/LUMO with their logs.